### 2019-08-31 #1692 @labarba
👋 @jrperilla — Thanks for agreeing to review this JOSS submission! I'm on the lookout for a second reviewer, and once we add them, we can start the review issue. In the meantime, feel free to start checking out the software, and you can also open issues in their repository as needed.
