### 2019-09-01 #1698 @whedon
[REVIEW]: Computing diffusion coeﬀicients in macromolecular simulations: the Diffusion Coeﬀicient Tool for VMD
**Submitting author:** @tonigi (<a href="http://orcid.org/0000-0001-6449-0596">Toni Giorgino</a>)
**Repository:** <a href="https://github.com/giorginolab/vmd_diffusion_coefficient" target ="_blank">https://github.com/giorginolab/vmd_diffusion_coefficient</a>
**Version:** v1.2
