### 2019-09-01 #1698 @labarba
👋  @jrperilla, @HaoZeke — Thank you for agreeing to review this submission for JOSS. At the top of this Review issue, each of you has a review checklist: work your way through it, and post any questions or comments for the authors here. You can also open new issues in the submission repository, linking to this thread here for a cross-reference. Godspeed!
