### 2019-09-03 #1698 @HaoZeke
Thanks @labarba. Right, so, off the bat, @tonigi, the `LICENSE` file has issues, essentially, it is currently not an [OSI approved license](https://opensource.org/licenses/alphabetical). I am opening an issue in the repo to discuss this further.
