### 2019-09-03 giorginolab/vmd_diffusion_coefficient#1 @HaoZeke
[JOSS] OSI License
Hi. This is linked to the [ongoing JOSS review](https://github.com/openjournals/joss-reviews/issues/1698#event-2599680575). Please replace the text of your license file with a suitable [OSI License](https://opensource.org/licenses/alphabetical). Personally, I think the [BSD 3-clause](https://opensource.org/licenses/BSD-3-Clause) is essentially what is described by the current `LICENSE` file, but feel free to choose another or discuss it further here.
