### 2019-09-03 giorginolab/vmd_diffusion_coefficient#1 @tonigi
Hi HaoZeke, thanks for the ongoing review! I updated the LICENSE file to the BSD-3, which is what was supposed to be all along.
