### 2019-09-03 #1698 @tonigi
Thanks for spotting! Updated LICENSE to the BSD-3 text.
