### 2019-09-03 giorginolab/vmd_diffusion_coefficient#1 @HaoZeke
Excellent! Thanks for the prompt resolution.
