### 2019-09-10 #1692 @labarba
👋  @jrperilla — the review is happening here: https://github.com/openjournals/joss-reviews/issues/1698
