### 2019-09-16 giorginolab/vmd_diffusion_coefficient#2 @HaoZeke
[JOSS] Doc Citation
Hi, as part of https://github.com/openjournals/joss-reviews/issues/1698, could you update the citation in the [readme](https://github.com/giorginolab/vmd_diffusion_coefficient/tree/master/doc).
