### 2019-09-16 giorginolab/vmd_diffusion_coefficient#3 @HaoZeke
[JOSS] Community Guidelines (Contribution)
Part of https://github.com/openjournals/joss-reviews/issues/1698. I think the docs are clear on (2) and (3), so if there could be a line added to address (1) then we're good to go!
