### 2019-09-16 f8097cb
Add explanation of fit button
