+++ doc/README.md
@@ -168,7 +168,7 @@ Citation
 
 Please cite the following publication:
 
-> Toni Giorgino, Computing diffusion coefficients in macromolecular simulations: the Diffusion Coefficient Tool for VMD, (2015). Available from [GitHub](https://github.com/giorginolab/vmd_diffusion_coefficient/).
+> Toni Giorgino, Computing diffusion coefficients in macromolecular simulations: the Diffusion Coefficient Tool for VMD, (2019). Journal of Open Source Software (under review). [doi:10.21105.joss.01698](http://doi.org/doi:10.21105.joss.01698) Available from [GitHub](https://github.com/giorginolab/vmd_diffusion_coefficient/). 
 
 
 
