### 2019-09-16 giorginolab/vmd_diffusion_coefficient#2 @tonigi
Done in 04d4ab981e5350207be828dae8aa076b5c10e303 - assuming the paper will be accepted :)
