### 2019-09-16 giorginolab/vmd_diffusion_coefficient#3 @tonigi
Done in 4587dd512b3a2e112eea57392bddbe85d6736b82  (together with an explanation of the test case and its comparision with the experimental result). Thanks!
