+++ doc/README.md
@@ -13,13 +13,14 @@ Installation
 VMD 1.9.2 or later is required. For installation instructions, please see [here](https://gist.github.com/tonigi/a9cfaf7642a7fbc13293).
 
 To test, see the [regtest.tcl](../demo/regtest.tcl) script in the
-`demo` directory. The test computes the lateral diffusion coefficient of
-cholesterol in a 1:1 POPC:CHL membrane on the basis of pre-computed
+`demo` directory. The test computes the lateral diffusion coefficient
+of cholesterol in a 1:1 POPC:CHL membrane on the basis of pre-computed
 simulation data from the *Case Study* supplement of [Guixà-González et
 al.](http://doi.org/10.1093/bioinformatics/btu037). The result, 0.55 ±
-0.031 Å²/ns (at 310 °K), matches well the experimental value 
-of 5.5 10<sup>-12</sup> m²/s (at 308 °K)
-in [Filippov et al.](https://doi.org/10.1021/la034222x).
+0.031 Å²/ns (at 310 °K), matches well the experimental value of 5.5
+10<sup>-12</sup> m²/s (at 308 °K) in Table 1 of [Filippov et
+al.](https://doi.org/10.1021/la034222x) (see below for units
+conversions).
 
 
 Usage (GUI)
