### 2019-09-16 #1698 @tonigi
@HaoZeke made amendments according to suggestions. Also, regarding the "Functionality" point, added an explanation of the demo calculation and how the results compare with experimental ones. :)
