### 2019-09-16 giorginolab/vmd_diffusion_coefficient#2 @HaoZeke
Excellent.
