### 2019-09-16 giorginolab/vmd_diffusion_coefficient#3 @HaoZeke
Thanks for the quick resolution!
