### 2019-09-16 #1698 @HaoZeke
@tonigi Thanks for the very prompt resolution of all points raised. @labarba I am pleased to say that I have completed my review, and recommend this for publication.
