+++ paper/paper.bib
@@ -90,3 +90,28 @@
 	pages = {2109--2117},
 	file = {Martin et al_2002_Apparent Subdiffusion Inherent to Single Particle Tracking.pdf:/Users/toni/Sync/literature/zotero/MD/Didactic/Martin et al_2002_Apparent Subdiffusion Inherent to Single Particle Tracking.pdf:application/pdf}
 }
+
+
+@article{guixa-gonzalez_membplugin:_2014,
+	title = {{MEMBPLUGIN}: studying membrane complexity in {VMD}},
+	volume = {30},
+	issn = {1367-4803, 1460-2059},
+	shorttitle = {{MEMBPLUGIN}},
+	url = {http://bioinformatics.oxfordjournals.org/content/30/10/1478},
+	doi = {10.1093/bioinformatics/btu037},
+	abstract = {Summary: Computer simulations are giving way to more complex and accurate studies of biological membranes by molecular dynamics (MD) simulations. The analysis of MD trajectories comprises the biophysical characterization of membrane properties or the study of protein–lipid interactions and dynamics. However, there is a lack of automated tools to analyse MD simulations of complex membrane or membrane-protein systems. Here we present MEMBPLUGIN, a plugin for the Visual Molecular Dynamics package that provides algorithms to measure a host of essential biophysical properties in simulated membranes. MEMBPLUGIN features are accessible both through a user-friendly graphical interface and as command-line procedures to be invoked in analysis scripts.
+Availability and implementation: MEMBPLUGIN is a VMD extension written in Tcl. Multi-platform source code, documentation and tutorials are freely available at http://membplugin.sourceforge.net.
+Contact: toni.giorgino@isib.cnr.it or jana.selent@upf.edu
+Supplementary information: Supplementary data are available at Bioinformatics online.},
+	language = {en},
+	number = {10},
+	urldate = {2014-05-19},
+	journal = {Bioinformatics},
+	author = {Guixà-González, Ramon and Rodriguez-Espigares, Ismael and Ramírez-Anguita, Juan Manuel and Carrió-Gaspar, Pau and Martinez-Seara, Hector and Giorgino, Toni and Selent, Jana},
+	month = may,
+	year = {2014},
+	pmid = {24451625},
+	pages = {1478--1480},
+	file = {Guixà-González et al_2014_MEMBPLUGIN.pdf:/Users/toni/Sync/literature/zotero/Toni/Toni_1_Journals/Guixà-González et al_2014_MEMBPLUGIN.pdf:application/pdf;Snapshot:/Users/toni/Zotero/storage/NTSIR699/1478.html:text/html;Snapshot:/Users/toni/Zotero/storage/Q7J25QFJ/bioinformatics.html:text/html}
+}
+
