### 2019-09-16 giorginolab/vmd_diffusion_coefficient#4 @jrperilla
Auto_path variable issue
When installing the plugin, the authors note (in the README of the doc/ subdirectory of the Git repository specified in the JOSS manuscript) several methods to integrate their plugin into the VMD startup routine. I opted to append the plugin location to the auto_path variable. However, when this is done, an error is thrown to stdout by VMD upon startup. While this did not hamper the functionality of the method, both in command-line or GUI usage (including registering the plugin in the Extensions menu), it indicates a bad interface with the VMD startup script/routines and the problem should be amended. An identical error is also thrown for VMD for Windows 10, which makes sense since Tcl/Tk-only plugins are architecture-independent.
