### 2019-09-16 giorginolab/vmd_diffusion_coefficient#4 @jrperilla
![image](https://user-images.githubusercontent.com/15744430/64998221-45995b00-d8b1-11e9-94e8-b30c858781b2.png)
