### 2019-09-16 giorginolab/vmd_diffusion_coefficient#5 @jrperilla
Crash during initialization
During testing, I attempted to analyze each water in every frame of the trajectory specified above, which totaled over 400,000 frames. VMD crashed during initialization of the ‘diffusion_coefficient_msd’ calculation. While this is a rather idiotic usage of the tool, it points to a problem with the method: that the initialization has no check for user input size or specified dimensionality. Given this tool has a GUI implementation, it may be smart to perform a simple check before array initialization to warn, or advise, users to prune their input or decrease the number of atoms in the selection to prevent a complete failure of VMD. Likewise, the initialization procedure seems agnostic to the dimensions specified by the user. Since diffusion along any permutation of x, y, or z may be analyzed at the user’s discretion, why does ‘diffusion_coefficient_msd’ always get and load x, y, z into arrays while the following call to ‘msd_between’ is the point at which the user’s arguments are checked?

![image](https://user-images.githubusercontent.com/15744430/64998235-50ec8680-d8b1-11e9-838b-261c98dcc6d5.png)
