### 2019-09-16 giorginolab/vmd_diffusion_coefficient#6 @jrperilla
calculating diffusion coefficient from linear fit
While using the GUI implementation, and selecting either of the two ‘Plot’ options, it would be beneficial to always output the diffusion coefficient from linear fit to stdout. Since the MSD calculation is expensive, I don’t see a reason not to perform the linear fit and extract the coefficient before calling the Multiplot window, even if the user did not click the separate “Diffusion coefficient from linear fit” button which launches the complete routine anyway but does not plot the MSD.
