### 2019-09-16 giorginolab/vmd_diffusion_coefficient#7 @jrperilla
Spelling issues
In the GUI window, as well as several other places in the documentation and in the JOSS manuscript, “specie” is written instead of “species.”
