### 2019-09-16 giorginolab/vmd_diffusion_coefficient#8 @jrperilla
Killswitch
Is it possible to add a “stop” or “abort” button to the GUI, to give users the chance to halt the calculation without having to kill VMD?
