### 2019-09-16 giorginolab/vmd_diffusion_coefficient#9 @jrperilla
Feature request
After instantiating the calculation with the GUI, it might be beneficial to print the equivalent command-line arguments to stdout, which will give users a better idea how to integrate the tool into their own scripts/analysis procedures.
