### 2019-09-16 giorginolab/vmd_diffusion_coefficient#10 @jrperilla
Printout functionality
It would be nice to add a button to the GUI to call the ‘printout’ procedure, in case a user wants to write data to a specified file without resorting to exporting the data from the Multiplot window.
