### 2019-09-16 #1698 @jrperilla
The Diffusion Coefficient Tool in VMD offers a simple and accessible method to glean diffusion coefficients from mean square displacement of a molecular species within a molecular dynamics trajectory. Given that diffusion coefficients computed from MD data can help to validate equilibration or otherwise detect the presence of bulk motions, this software is of particular interest to the computational chemistry community. To test the method appropriately, i.e., following guidelines stated by the author in the JOSS manuscript, an equilibrated system was chosen describing roughly 30,000 TIP3P waters over 4 microseconds. While the plugin functioned as intended and produced consistent results given different values of stride and lag intervals, from both command-line and GUI, several concerns regarding the structure and design of the tool were noted during testing.

1. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/4
2. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/5
3. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/6

Minor issues or feature requests:

1. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/7
2. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/8
3. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/9
4. https://github.com/giorginolab/vmd_diffusion_coefficient/issues/10
