### 2019-09-18 ec8a675
memory-efficient computation
