+++ doc/README.md
@@ -12,6 +12,11 @@ Installation
 
 VMD 1.9.2 or later is required. For installation instructions, please see [here](https://gist.github.com/tonigi/a9cfaf7642a7fbc13293).
 
+The plugin tries (mis)uses TCL's `auto_path` mechanism to register
+itself with VMD's Extensions menu. This does not work in all
+environments. In case of problems, the `load.tcl` file lists the
+required steps explicitly.
+
 
 Test and example
 ----------------
