### 2019-09-18 giorginolab/vmd_diffusion_coefficient#4 @tonigi
The script used a trick in `pkgIndex.tcl` to register itself with VMD; it was supposed to enable loading by merely setting TCLLIBPATH. However this did not work in all situations (as in your case).   I rolled back the install instructions to a more standard "append auto_path / package require / vmd register" sequence, now in `load.tcl`. b93901dc72982373faecea357fb85b428d736d24
