### 2019-09-18 giorginolab/vmd_diffusion_coefficient#5 @tonigi
Re-wrote the algorithm to avoid the coordinate copy which occurred in the initialization. The calculation should now take a negligible amount of memory. ec8a675534f1f977ab90d051fa1d2b403343a13d
