### 2019-09-18 giorginolab/vmd_diffusion_coefficient#6 @tonigi
Indeed the button was more a "cosmetic" choice to highlight the fact that a fit was taking place. Good point though. I changed the GUI according to the suggestion - the button is gone; D from fit is now implicitly printed and added to the plots. 46a63521cb291a130b32972ddf43082ba316420e
