### 2019-09-18 giorginolab/vmd_diffusion_coefficient#7 @tonigi
I naively followed the latin declination for species. This is now corrected. 0fad1a2e6bdc5a90bf34e7f47acedaa04af076be
