### 2019-09-18 giorginolab/vmd_diffusion_coefficient#8 @tonigi
Implemented in cfe9afbdf887eba742a824f968f51b2066bfb046, thanks
