### 2019-09-18 giorginolab/vmd_diffusion_coefficient#9 @tonigi
Good idea. Implemented in 3cc46060f5c8b5eacd0ebbbec7f6805cdc5c34a4
