### 2019-09-18 giorginolab/vmd_diffusion_coefficient#10 @tonigi
Done in 159ff92f7229eb1aa2c39a8aeb3e480bf9b32e45 et al. , as a "save" button. Saves a CSV with tau/MSD/D and headers to reproduce the computation.
