### 2019-09-18 #1698 @tonigi
@jrperilla Thanks for the suggestions! Each item is answered in the corresponding issue. In short, all the added functionality now constitutes release 1.2. The paper and docs are updated accordingly.
