### 2019-09-18 #1698 @jrperilla
I think all my concerns were addressed and the paper is ready for publication.
