### 2019-09-18 #1698 @jrperilla
I want to shout out to @abryer94 for helping with the testing of the code.
