+++ README.md
@@ -5,9 +5,11 @@ Toni Giorgino © 2011-2019
 
 The Diffusion Coefficient Tool is an analysis plugin for VMD that
 computes one, two or three-dimensional MSD-based diffusion
-coefficients of a chosen molecular species.
+coefficients of a chosen molecular species. 
 
 Please find this software's documentation in the [doc subdirectory](doc).
+
+[![status](https://joss.theoj.org/papers/25d51faf73cc17ae3affb51b787bbe18/status.svg)](https://joss.theoj.org/papers/25d51faf73cc17ae3affb51b787bbe18)
 
 License
 -------
