### 2019-09-18 #1698 @labarba
@tonigi — We're now going to need you to make an archive deposit in Zenodo (or a similar service) get a DOI, and post the DOI here.

Please edit the archive metadata so the title and author list matches the JOSS paper.
