### 2019-09-18 #1698 @tonigi
Hi @labarba , just tried with Zenodo's GitHub integration and got ...

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3445530.svg)](https://doi.org/10.5281/zenodo.3445530)

10.5281/zenodo.3445530
