### 2019-09-19 #1698 @tonigi
@HaoZeke , @jrperilla , @abryer94 -- thanks for the thorough reviews!

@labarba -- thanks for managing!

Sorry for the nitpicking - I just wanted to point out that non-ascii "ff" combined characters slipped in the title. I guess it's better to rename them to plain "ff" before submission lest we hit some "undocumented feature" in the next steps such as cross-ref indexing etc.
