### 2019-09-19 #1698 @labarba
Congratulations, @tonigi, your JOSS paper is now published! 🚀

Huge thanks to our reviewers: @jrperilla, @HaoZeke — your contribution to JOSS is very much appreciated! 🙏
